assist: fix exports in the Prolog ontology generation
The original code was producing incorrect results when there were no
properties, as the last thing would have a trailing comma, not followed by any
property.

This improved version first aggregates together all things and properties
strings, then uses a combinator to insert commas as appropriate.

For an example of where this went wrong, try and generate software.pl with the
current state of the ontology.